requirements.txt
fastapi
ollama
llama-index-llms-ollama
# langchain
# llama-index-llms-openai
# llama-index-program-openai